LayoutTests/http/tests/security/mixedContent/insecure-css-in-iframe.html

<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

window.data = null;
window.addEventListener("message", function (e) {
    window.data = e.data;
    shouldBe("window.data", "'blocked'");

    if (window.testRunner)
        testRunner.notifyDone();
}, false);
</script>
<p>This test loads a secure iframe that loads an insecure style sheet.
We should trigger a mixed content callback because an active network attacker
can use CSS3 to breach the confidentiality of the HTTPS security origin.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css.html";
></iframe>
</body>
</html>
